+++ addons/br/nfse/extensions.go
@@ -12,14 +12,33 @@ import (
 // of these extensions are common, they can be moved to the regime or to a
 // shared addon.
 const (
+	ExtKeyCNAE            = "br-nfse-cnae"
 	ExtKeyFiscalIncentive = "br-nfse-fiscal-incentive"
+	ExtKeyISSLiability    = "br-nfse-iss-liability"
 	ExtKeyMunicipality    = "br-nfse-municipality"
 	ExtKeyService         = "br-nfse-service"
-	ExtKeySimplesNacional = "br-nfse-simples-nacional"
+	ExtKeySimples         = "br-nfse-simples"
 	ExtKeySpecialRegime   = "br-nfse-special-regime"
 )
 
 var extensions = []*cbc.KeyDefinition{
+	{
+		Key: ExtKeyCNAE,
+		Name: i18n.String{
+			i18n.EN: "CNAE code",
+			i18n.PT: "Código CNAE",
+		},
+		Desc: i18n.String{
+			i18n.EN: here.Doc(`
+				The CNAE (National Classification of Economic Activities) code for a service.
+
+				List of codes from the IBGE (Brazilian Institute of Geography and Statistics):
+
+				* https://www.ibge.gov.br/en/statistics/technical-documents/statistical-lists-and-classifications/17245-national-classification-of-economic-activities.html
+			`),
+		},
+		Pattern: `^\d{2}[\s\.\-\/]?\d{2}[\s\.\-\/]?\d[\s\.\-\/]?\d{2}$`,
+	},
 	{
 		Key: ExtKeyFiscalIncentive,
 		Name: i18n.String{
@@ -46,9 +65,78 @@ var extensions = []*cbc.KeyDefinition{
 			i18n.EN: here.Doc(`
 				Indicates whether a party benefits from a fiscal incentive.
 
-				List of codes taken from the national NFSe standard:
-				https://abrasf.org.br/biblioteca/arquivos-publicos/nfs-e-manual-de-orientacao-do-contribuinte-2-04/download
+				List of codes from the national NFSe ABRASF (v2.04) model:
+
+				* https://abrasf.org.br/biblioteca/arquivos-publicos/nfs-e-manual-de-orientacao-do-contribuinte-2-04/download
 				(Section 10.2, Field B-68)
+			`),
+		},
+	},
+	{
+		Key: ExtKeyISSLiability,
+		Name: i18n.String{
+			i18n.EN: "ISS Liability",
+			i18n.PT: "Exigibilidade ISS",
+		},
+		Values: []*cbc.ValueDefinition{
+			{
+				Value: "1",
+				Name: i18n.String{
+					i18n.EN: "Liable",
+					i18n.PT: "Exigível",
+				},
+			},
+			{
+				Value: "2",
+				Name: i18n.String{
+					i18n.EN: "Not subject",
+					i18n.PT: "Não incidência",
+				},
+			},
+			{
+				Value: "3",
+				Name: i18n.String{
+					i18n.EN: "Exempt",
+					i18n.PT: "Isenção",
+				},
+			},
+			{
+				Value: "4",
+				Name: i18n.String{
+					i18n.EN: "Export",
+					i18n.PT: "Exportação",
+				},
+			},
+			{
+				Value: "5",
+				Name: i18n.String{
+					i18n.EN: "Immune",
+					i18n.PT: "Imunidade",
+				},
+			},
+			{
+				Value: "6",
+				Name: i18n.String{
+					i18n.EN: "Suspended Judicially",
+					i18n.PT: "Suspensa por Decisão Judicial",
+				},
+			},
+			{
+				Value: "7",
+				Name: i18n.String{
+					i18n.EN: "Suspended Administratively",
+					i18n.PT: "Suspensa por Processo Administrativo",
+				},
+			},
+		},
+		Desc: i18n.String{
+			i18n.EN: here.Doc(`
+				Indicates the ISS liability status, i.e., whether the ISS tax is due or not and why.
+
+				List of codes from the national NFSe ABRASF (v2.04) model:
+
+				* https://abrasf.org.br/biblioteca/arquivos-publicos/nfs-e-manual-de-orientacao-do-contribuinte-2-04/download
+				(Section 10.2, Field B-38)
 			`),
 		},
 	},
@@ -63,7 +151,7 @@ var extensions = []*cbc.KeyDefinition{
 				The municipality code as defined by the IGBE (Brazilian Institute of Geography and
 				Statistics).
 
-				For further details on the list of possible codes, see:
+				List of codes from the IGBE:
 
 				* https://www.ibge.gov.br/explica/codigos-dos-municipios.php
 			`),
@@ -88,9 +176,9 @@ var extensions = []*cbc.KeyDefinition{
 		},
 	},
 	{
-		Key: ExtKeySimplesNacional,
+		Key: ExtKeySimples,
 		Name: i18n.String{
-			i18n.EN: "Opting for “Simples Nacional”",
+			i18n.EN: "Opting for “Simples Nacional” regime",
 			i18n.PT: "Optante pelo Simples Nacional",
 		},
 		Values: []*cbc.ValueDefinition{
@@ -111,10 +199,13 @@ var extensions = []*cbc.KeyDefinition{
 		},
 		Desc: i18n.String{
 			i18n.EN: here.Doc(`
-				Indicates whether a party is opting for the “Simples Nacional” tax regime.
+				Indicates whether a party is opting for the “Simples Nacional” (Regime Especial
+				Unificado de Arrecadação de Tributos e Contribuições devidos pelas Microempresas e
+				Empresas de Pequeno Porte) tax regime
 
-				List of codes taken from the national NFSe standard:
-				https://abrasf.org.br/biblioteca/arquivos-publicos/nfs-e-manual-de-orientacao-do-contribuinte-2-04/download
+				List of codes from the national NFSe ABRASF (v2.04) model:
+
+				* https://abrasf.org.br/biblioteca/arquivos-publicos/nfs-e-manual-de-orientacao-do-contribuinte-2-04/download
 				(Section 10.2, Field B-67)
 			`),
 		},
@@ -171,10 +262,11 @@ var extensions = []*cbc.KeyDefinition{
 		},
 		Desc: i18n.String{
 			i18n.EN: here.Doc(`
-				Indicates a special tax regime that the party is subject to.
+				Indicates a special tax regime that a party is subject to.
 
-				List of codes taken from the national NFSe standard:
-				https://abrasf.org.br/biblioteca/arquivos-publicos/nfs-e-manual-de-orientacao-do-contribuinte-2-04/download
+				List of codes from the national NFSe ABRASF (v2.04) model:
+
+				* https://abrasf.org.br/biblioteca/arquivos-publicos/nfs-e-manual-de-orientacao-do-contribuinte-2-04/download
 				(Section 10.2, Field B-66)
 			`),
 		},