+++ CHANGELOG.md
@@ -10,6 +10,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - ae: added UAE regime
 
+### Changed
+
+- `es-tbai-v1`: always add `es-tbai-product` extension to items, indicating default value.
+
+### Fixed
+
+- `es-tbai-v1`: issue with preceding validation.
+
 ## [v0.205.1] - 2024-11-19
 
 ### Added