+++ CHANGELOG.md
@@ -8,14 +8,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
-- ae: added UAE regime
+- `ae`: added UAE regime
+- `br`: supplier extensions, validations & identities
 
 ## [v0.205.1] - 2024-11-19
 
 ### Added
 
 - `org`: `Address` includes `LineOne()`, `LineTwo()`, `CompleteNumber()` methods to help with conversion to other formats with some regional formatting.
-- `br`: supplier extensions, validations & identities
 
 ### Changes